Revert "[EMISWEB-142] delay when export load is at 90% solved"
